broken impl

+++ src/listings/ch06.rs
@@ -682,7 +682,7 @@ pub fn calc_accuracy_loader(
             while let Some(Ok((input_batch, target_batch))) = data_batcher.next() {
                 let num_correct =
                     calc_num_correct_batch(&input_batch, &target_batch, model, device)?;
-                correct_predictions += num_correct.to_scalar::<u32>()? as usize;
+                correct_predictions += num_correct.to_scalar::<u8>()? as usize;
                 num_examples += input_batch.dims()[0];
             }
             Ok(correct_predictions as f32 / num_examples as f32)
@@ -691,7 +691,7 @@ pub fn calc_accuracy_loader(
             while let Some(Ok((input_batch, target_batch))) = data_batcher.next() {
                 let num_correct =
                     calc_num_correct_batch(&input_batch, &target_batch, model, device)?;
-                correct_predictions += num_correct.to_scalar::<u32>()? as usize;
+                correct_predictions += num_correct.to_scalar::<u8>()? as usize;
                 num_examples += input_batch.dims()[0];
                 if num_examples >= n {
                     break;
